requirements.txt
flask 
av 
pillow==9.0.1
